clean : clear unnecessary file
