+++ app/src/main/java/com/android/shelfLife/ui/navigation/TopNavigationMenu.kt
@@ -5,13 +5,18 @@ import androidx.compose.animation.expandVertically
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.shrinkVertically
+import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Check
 import androidx.compose.material.icons.filled.FilterList
 import androidx.compose.material.icons.filled.Menu
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.FilterChip
+import androidx.compose.material3.FilterChipDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
@@ -22,6 +27,7 @@ import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
@@ -32,7 +38,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.android.shelfLife.model.household.HouseHold
 import com.android.shelfLife.model.household.HouseholdViewModel
-import com.android.shelfLife.ui.utils.FilterBar
+import com.android.shelfLife.ui.navigation.FilterBar
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -109,4 +115,53 @@ fun HouseHoldElement(
       selected = household == selectedHousehold,
       onClick = { onHouseholdSelected(household) },
       modifier = Modifier.padding(NavigationDrawerItemDefaults.ItemPadding))
+}
+
+@Composable
+fun FilterChipItem(text: String, isSelected: Boolean, onClick: () -> Unit) {
+    FilterChip(
+        selected = isSelected,
+        onClick = onClick,
+        label = { Text(text = text) },
+        leadingIcon =
+        if (isSelected) {
+            { Icon(imageVector = Icons.Default.Check, contentDescription = "Selected") }
+        } else null,
+        colors =
+        FilterChipDefaults.filterChipColors(
+            selectedContainerColor = MaterialTheme.colorScheme.secondary,
+            selectedLabelColor = Color.White,
+            selectedLeadingIconColor = Color.White,
+            containerColor = Color.White,
+            labelColor = Color.Black
+        ),
+        modifier = Modifier.padding(horizontal = 4.dp, vertical = 8.dp) // Add padding between chips
+    )
+}
+
+@Composable
+fun FilterBar(filters: List<String>) {
+  // State to track the selection of each filter chip
+  val selectedFilters = remember { mutableStateListOf<String>() }
+  val scrollState = rememberScrollState()
+
+    Row(
+        modifier =
+        Modifier.horizontalScroll(scrollState) // Enables horizontal scrolling
+            .padding(horizontal = 8.dp, vertical = 4.dp)
+    ) {
+        filters.forEach { filter ->
+            val isSelected = selectedFilters.contains(filter)
+            FilterChipItem(
+                text = filter,
+                isSelected = isSelected,
+                onClick = {
+                    if (isSelected) {
+                        selectedFilters.remove(filter)
+                    } else {
+                        selectedFilters.add(filter)
+                    }
+                })
+        }
+    }
 }
